Separate game files
